+++ includes/modules/shipping/chronopost.php
@@ -119,7 +119,7 @@
       $cost = -1;
       for ($i = 0, $n = sizeof($table); $i < $n; $i+=2) {
         if ($shipping_weight <= $table[$i]) {
-          $cost = (double)$table[$i+1] + (double)MODULE_SHIPPING_CHRONOPOST_HANDLING;
+          $cost = (float)$table[$i+1] + (float)MODULE_SHIPPING_CHRONOPOST_HANDLING;
           break;
         }
       }
